+++ mrmustard/lab/states/base.py
@@ -405,13 +405,14 @@ class State(CircuitComponent):
         Returns:
             The Fock distribution.
         """
+        batch_shape = self.ansatz.batch_shape
         fock_array = self.fock_array(cutoff)
         if not self.wires.ket or not self.wires.bra:
-            return math.abs(fock_array) ** 2
-        delta = -self.n_modes - 1
+            return math.reshape(math.abs(fock_array) ** 2, (*batch_shape, -1))
+        axis1 = -self.n_modes - 1
         for i in range(self.n_modes):
-            fock_array = np.diagonal(fock_array, axis1=delta, axis2=-1 - i)
-        return math.abs(fock_array)
+            fock_array = math.diagonal(fock_array, axis1=axis1, axis2=-(1 + i))
+        return math.reshape(math.abs(fock_array), (*batch_shape, -1))
 
     @abstractmethod
     def formal_stellar_decomposition(