+++ mrmustard/math/backend_manager.py
@@ -491,6 +491,21 @@ class BackendManager:
         """
         return self._apply("det", (matrix,))
 
+    def diagonal(self, array: Tensor, offset: int = 0, axis1: int = 0, axis2: int = 0) -> Tensor:
+        r"""
+        Return specified diagonals of array.
+
+        Args:
+            array: The array to take the diagonal of.
+            offset: The offset of the diagonal.
+            axis1: The first axis to take the diagonal of.
+            axis2: The second axis to take the diagonal of.
+
+        Returns:
+            The diagonal of ``array``.
+        """
+        return self._apply("diagonal", (array, offset, axis1, axis2))
+
     def diag(self, array: Tensor, k: int = 0) -> Tensor:
         r"""The array made by inserting the given array along the :math:`k`-th diagonal.