+++ .github/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Added a ``rich`` based repr to ``ParameterSet``.
 [(#616)](https://github.com/XanaduAI/MrMustard/pull/616)
 
+### Bug fixes
+* Fixed a bug with ``State.fock_distribution`` where batch dimensions and mult-mode states were not handled correctly.
+[(#635)](https://github.com/XanaduAI/MrMustard/pull/635)
+
 ---
 
 # Release 1.0.0a1